KAFKA-654 The value of property AUTO_COMMIT_INTERVAL_MS_CONFIG has Long type

+++ clients/src/main/java/org/apache/kafka/clients/consumer/KafkaConsumer.java
@@ -880,7 +880,7 @@ public class KafkaConsumer<K, V> implements Consumer<K, V> {
                                 metricGrpPrefix,
                                 this.time,
                                 enableAutoCommit,
-                                config.getInt(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG),
+                                config.getLong(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG),
                                 this.interceptors,
                                 config.getBoolean(ConsumerConfig.THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED));
                     this.fetcher = new Fetcher<>(
